export preview lfg
