+++ apps/dotcom/client/src/tla/components/TlaFileShareMenu/file-share-menu.module.css
@@ -75,26 +75,7 @@
 	right: 10px;
 }
 
-/* QR Code */
-
-.qrCode {
-	position: relative;
-	width: 100%;
-	height: 216px;
-	padding: 0px var(--share-menu-padding);
-	margin: 0px 0px var(--share-menu-padding) 0px;
-}
-
-.qrCodeInner {
-	width: 100%;
-	height: 100%;
-	background-color: var(--tla-color-low);
-	background-size: 100% 100%;
-	background-position: center center;
-	background-color: var(--tla-color-low);
-}
-
-/* Share menu description */
+/* Help item */
 
 .helpItem {
 	position: relative;
@@ -113,4 +94,56 @@
 .helpItem p {
 	margin: 0px;
 	padding: 12px;
+}
+
+/* QR Code */
+
+.qrCode {
+	position: relative;
+	width: 100%;
+	height: 216px;
+	overflow: hidden;
+	border-radius: 4px;
+	padding: 8px;
+	border: 1px solid var(--tla-color-border);
+	background-color: var(--tla-color-panel);
+}
+
+.qrCodeInner {
+	width: 100%;
+	height: auto;
+	border-radius: 3px;
+	pointer-events: all;
+}
+
+/* Export preview */
+
+.exportPreview {
+	position: relative;
+	width: 100%;
+	height: 216px;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	overflow: hidden;
+	border-radius: 4px;
+	border: 1px solid var(--tla-color-border);
+	background-color: var(--tla-color-panel);
+}
+
+.exportPreviewInner {
+	width: auto;
+	height: auto;
+	max-height: 216px;
+	pointer-events: all;
+	border: 1px solid var(--tla-color-border);
+	box-sizing: content-box;
+}
+
+.exportPreviewSize {
+	position: absolute;
+	bottom: 0px;
+	right: 0px;
+	background-color: var(--tla-color-panel);
+	padding: 2px 4px;
 }
