+++ apps/dotcom/client/src/tla/utils/getCurrentEditor.ts
@@ -1,0 +1,5 @@
+import { Editor } from 'tldraw'
+
+export function getCurrentEditor() {
+	return (window as any).editor as Editor | undefined
+}
