+++ apps/dotcom/client/src/utils/globalEditor.ts
@@ -1,0 +1,3 @@
+import { atom, Editor } from 'tldraw'
+
+export const globalEditor = atom<Editor | null>('globalEditor', null)
